+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Add cleanup call on every request and by configured interval
+
 ### Changed
 - Prevent encryption animation to take too long
 
 ### Removed
 - Remove legacy appearance.css
+- Remove CLI call for cleanup
 
 ### Fixed
 - Fix condition race for activeuntil on show message page